Generalised CLI flags overriding hardcoded values
